rename BasketActivity and BuyProductsActivity
