+++ Seed/app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".BasketActivity"></activity>
+        <activity android:name=".data.BuyProductsActivity"></activity>
         <activity android:name=".SearchStoreLocation" />
         <activity android:name=".BuyProductsActivity" />
         <activity android:name=".SearchMyLocation" />
