+++ Seed/app/src/main/java/com/example/seed/BuyProductsActivity.java
@@ -1,21 +1,6 @@
 package com.example.seed;
 
-import androidx.appcompat.app.AppCompatActivity;
+import android.app.Activity;
 
-import android.os.Bundle;
-import android.widget.TextView;
-
-import org.w3c.dom.Text;
-
-import java.util.Date;
-
-public class BuyProductsActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_buy_products);
-
-    }
-
+public class BuyProductsActivity extends Activity {
 }
