+++ Seed/app/src/main/java/com/example/seed/data/BuyProductsActivity.java
@@ -1,16 +1,19 @@
-package com.example.seed;
+package com.example.seed.data;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class BasketActivity extends AppCompatActivity {
+import com.example.seed.R;
+
+public class BuyProductsActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_basket);
+        setContentView(R.layout.activity_buy_products);
+
+
 
     }
-
 }
